refactor: 拆分 Gateway 大文件模块
